removing the misc plugin and placing its routes in the appropriate plugin
