+++ test/back/endpoints/runner.js
@@ -1,5 +1,4 @@
 require('./modules.test.js');
 require('./users.test.js');
 require('./quizes.test.js');
-require('./misc.test.js');
 require('./index.test.js');
